Move Event type to channels package
I dislike this, because an event is philosophically unconnected to a
channel, but a generic "event" thing. But, the event really needs a
reference to the channel, so that it can send replies back through it.
Go's absolute refusal of circular imports means that this basically has
to be in the channels package.